requirements.txt
cairocffi==0.8.0
CairoSVG==2.1.1
certifi==2018.4.16
cffi==1.11.2
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
dj-stripe==1.0.1
Django==1.11.7
django-filter==1.1.0
django-polymorphic==2.0.2
django-rest-auth==0.9.3
djangorestframework==3.7.7
html5lib==0.999999999
idna==2.6
jsonfield==2.0.2
Markdown==2.6.11
mysqlclient==1.3.13
olefile==0.44
pdfrw==0.4
Pillow==4.3.0
pycparser==2.18
Pyphen==0.9.4
python-dateutil==2.6.1
python-doc-inherit==0.3.0
python-magic==0.4.13
pytz==2017.3
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
stripe==1.82.1
tinycss2==0.6.1
tqdm==4.23.4
urllib3==1.22
WeasyPrint==0.41
webencodings==0.5.1
